product Explore +product comment adapter ans list +product owner info +product  rate

+++ app/src/main/java/com/ronaker/app/data/DefaultOrderRepository.kt
@@ -76,7 +76,7 @@ class DefaultOrderRepository(private val api: OrderApi) :
         token: String?,
         orderSuid: String,
         comment: String,
-        stars: Int
+        stars: Double
     ): Observable<Result<Boolean>> {
 
         val request=
