+++ app/src/main/java/com/ronaker/app/data/network/response/ProductItemResponceModel.kt
@@ -16,7 +16,7 @@ data class ProductItemResponceModel(
     @SerializedName("address")  val address: String?,
     @SerializedName("review_status")  val review_status: String?,
     @SerializedName("user_status")  val user_status: String?,
-    @SerializedName("rating")  val rating: Int?
+    @SerializedName("rating")  val rating: Double?
