+++ app/src/main/java/com/ronaker/app/data/network/response/ProductRatingResponceModel.kt
@@ -3,7 +3,7 @@ package com.ronaker.app.data.network.response
 import com.google.gson.annotations.SerializedName
 
 data class ProductRatingResponceModel(
-    @SerializedName("stars") val stars: Int,
+    @SerializedName("stars") val stars: Float,
     @SerializedName("comment") val comment: String
 ) {
